Extract inference objective assembly and share count-active indicators.
Centralize CP-SAT objective terms in inference_objective.py and reuse
add_count_active_indicator from constraints and ranking bin wiring.
